+++ flexmeasures/api/v3_0/assets.py
@@ -1682,7 +1682,8 @@ class AssetAPI(FlaskView):
                           flex-context:
                             consumption-price: {sensor: 9}
                             production-price: {sensor: 10}
-                            inflexible-device-sensors: [13, 14, 15]
+                            inflexible-consumption: [{sensor: 13}, {sensor: 14}]
+                            inflexible-production: [{sensor: 15}]
                             site-power-capacity: 100 kVA
                             site-production-capacity: 80 kW
                             site-consumption-capacity: {sensor: 32}
@@ -1720,7 +1721,8 @@ class AssetAPI(FlaskView):
                           flex-context:
                             consumption-price: {sensor: 9}
                             production-price: {sensor: 10}
-                            inflexible-device-sensors: [13, 14, 15]
+                            inflexible-consumption: [{sensor: 13}, {sensor: 14}]
+                            inflexible-production: [{sensor: 15}]
                             site-power-capacity: 100 kVA
                             site-production-capacity: 80 kW
                             site-consumption-capacity: {sensor: 32}